+++ src/libafb/wsapi/afb-wsapi.c
@@ -293,7 +293,7 @@ static int writebuf_putbuf(struct writebuf *wb, const void *value, int length)
 
 	/* check enough length */
 	nafter = n + length;
-	if (nafter > WRITEBUF_BUFSZ)
+	if (nafter > (int)WRITEBUF_BUFSZ)
 		return 0;
 
 	/* get where to store */
